Introduction of "MultitonBase" class
Introduction of "MultitonBase" class as a cleaner way of avoiding reinstantiations with the same arguments of the "AragogRunner" class. Also added a "_summarize_for_hash" method to deal with large Numpy arrays. The latter could take a lot of processing time, that is why we chose to determine a hash based on some average statistics. That may result in rare reinitializations of the "AragogRunner" class when in fact different arguments are applied, but will have a negligible effect on run times, generally speaking.