Pick up Julia install path
We observed, when issuing an install from scratch on Habrok, using "proteus install-all --export-env", that Julia was not found, so AGNI could not be installed.
This also happened when a default location was chosen, which is ~/.juliaup/bin.
So what needed to be collected from the "curl -fsSL https://install.julialang.org | sh" Julia install is an install path: install_julia --> install_julia_and_get_bin_path
I.e. the Julia install path should also be returned and added to "env" such that AGNI can be installed.
It is assumed that that install path is returned in the Julia install process, i.e. becomes part of stdout such that these lines will work:
"
        proc = subprocess.run(
            ["bash", "-c", "curl -fsSL https://install.julialang.org | sh"],
            check=True,
            capture_output=True,
            text=True,
        )
        # Search for the line that shows the install path
        match = re.search(r"installed at: (.+/bin)", proc.stdout)
"
Fingers crossed.